feat: support template cmd to generate topology files
