+++ cmd/cluster/start2.go
@@ -15,46 +15,56 @@
 package cluster
 
 import (
+	"fmt"
+
+	"github.com/openGemini/gemix/pkg/cluster/manager"
 	"github.com/openGemini/gemix/utils"
 	"github.com/spf13/cobra"
 )
 
-var startOpts utils.StartOptions
+// startCmd represents the start command
+var startCmd2 = &cobra.Command{
+	Use:   "start2",
+	Short: "Start an openGemini cluster",
+	Long:  `Start an openGemini cluster`,
+	Run: func(cmd *cobra.Command, args []string) {
+		var ops utils.ClusterOptions
+		var err error
+		if ops, err = ReadClusterOptionsByName(cmd); err != nil {
+			fmt.Println(err)
+			fmt.Println(cmd.UsageString())
+			return
+		}
 
-func startCmd2() *cobra.Command {
-	var (
-		initPasswd bool
-	)
+		err = StartCluster(ops)
+		if err != nil {
+			fmt.Println(err)
+			return
+		}
 
-	var cmd = &cobra.Command{
-		Use:   "start2 <cluster-name>",
-		Short: "Start an openGemini cluster",
-		Long:  `Start an openGemini cluster`,
-		RunE: func(cmd *cobra.Command, args []string) error {
-			if len(args) != 1 {
-				return cmd.Help()
-			}
+		fmt.Printf("\nCheck the status of openGemini cluster\n")
+		err = PatrolCluster(ops)
+		if err != nil {
+			fmt.Println(err)
+		}
+	},
+}
 
-			clusterName := args[0]
-			err := cm.StartCluster(clusterName, gOpt)
-			if err != nil {
-				return err
-			}
+func StartCluster(clusterOpts utils.ClusterOptions) error {
+	starter := manager.NewGeminiStarter(clusterOpts, startOpts)
+	defer starter.Close()
 
-			// TODO: init password
-			return nil
-		},
-		ValidArgsFunction: func(cmd *cobra.Command, args []string, toComplete string) ([]string, cobra.ShellCompDirective) {
-			switch len(args) {
-			case 0:
-				return shellCompGetClusterName(cm, toComplete)
-			default:
-				return nil, cobra.ShellCompDirectiveNoFileComp
-			}
-		},
+	if err := starter.PrepareForStart(); err != nil {
+		return err
 	}
-	cmd.Flags().BoolVar(&initPasswd, "init", false, "Initialize a secure root password for the database")
-	cmd.Flags().StringSliceVarP(&gOpt.Roles, "role", "R", nil, "Only start specified roles")
-	cmd.Flags().StringSliceVarP(&gOpt.Nodes, "node", "N", nil, "Only start specified nodes")
-	return cmd
+	if err := starter.Start(); err != nil {
+		return err
+	}
+	fmt.Printf("Successfully started the openGemini cluster with version : %s\n", clusterOpts.Version)
+	return nil
+}
+
+func init() {
+	startCmd2.Flags().StringP("name", "n", "", "cluster name")
+	startCmd2.Flags().BoolVarP(&startOpts.SkipCreateUser, "skip-create-user", "", false, "(EXPERIMENTAL) Skip creating the user specified in topology.")
 }
