+++ embed/examples/cluster/minimal.yaml
