+++ pkg/cluster/task/mkdir.go
@@ -46,9 +46,9 @@ func (m *Mkdir) Execute(ctx context.Context) error {
 		xs := strings.Split(dir, "/")
 		// Create directories recursively
 		// The directory /a/b/c will be flatten to:
-		// 		test -d /a || (mkdir /a && chown tidb:tidb /a)
-		//		test -d /a/b || (mkdir /a/b && chown tidb:tidb /a/b)
-		//		test -d /a/b/c || (mkdir /a/b/c && chown tidb:tidb /a/b/c)
+		// 		test -d /a || (mkdir /a && chown gemini:gemini /a)
+		//		test -d /a/b || (mkdir /a/b && chown gemini:gemini /a/b)
+		//		test -d /a/b/c || (mkdir /a/b/c && chown gemini:gemini /a/b/c)
 		for i := 0; i < len(xs); i++ {
 			if xs[i] == "" {
 				continue
